+++ src/component/atom/Text.js
